% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCohorts.R
\name{generateCohorts}
\alias{generateCohorts}
\title{generateCohorts}
\usage{
generateCohorts(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTableName,
  tempEmulationSchema,
  cohortsFolder = system.file(package = "ONLTreatmentPatternsBreastCancer", "cohorts")
)
}
\arguments{
\item{connectionDetails}{(`ConnectionDetails`)}

\item{cdmDatabaseSchema}{(`character(1)`)}

\item{cohortDatabaseSchema}{(`character(1)`)}

\item{cohortTableName}{(`character(1)`)}

\item{cohortsFolder}{(`character(1)`)}
}
\value{
(`list`)
}
\description{
generateCohorts
}
