% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTreatmentPatterns.R
\name{runTreatmentPatterns}
\alias{runTreatmentPatterns}
\title{runTreatmentPatterns}
\usage{
runTreatmentPatterns(
  connectinoDetails,
  resultSchema,
  cdmSchema,
  tempEmulationSchema,
  cohortTableName,
  outputFolder
)
}
\arguments{
\item{connectinoDetails}{(`ConnectionDetails`) ConnectionDetails object created with `DatabaseConnector::createConnectionDetails`.}

\item{resultSchema}{(`character(1)`) Database Schema where the cohort table exists.}

\item{cdmSchema}{(`character(1)`) Database schema wherethe CDM exists.}

\item{tempEmulationSchema}{(`character(1)`) Temporary emulation schema as described in `TreatmentPatterns::computePathways`.}

\item{cohortTableName}{(`character(1)`) Name of the cohort table.}

\item{outputFolder}{(`character(1)`) Path where the output should be written to.}
}
\value{
`NULL`
}
\description{
runTreatmentPatterns
}
